+++ client/python/datajunction/__init__.py
@@ -3,7 +3,7 @@ A DataJunction client for connecting to a DataJunction server
 """
 __version__ = "0.0.1a1"
 
-from datajunction.djclient import (
+from datajunction.client import (
     Cube,
     Dimension,
     DJClient,
